+++ client/src/lib.rs
@@ -3,6 +3,7 @@ pub mod editor;
 pub mod errors;
 
 use btep::{prelude::S2C, Deserialize};
+use editor::App;
 use core::panic;
 use crossterm::{
     event::{EnableBracketedPaste, Event, EventStream},
@@ -12,7 +13,6 @@ use crossterm::{
         self, disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen,
     },
 };
-use editor::Client;
 use futures::{future, FutureExt, StreamExt};
 use std::{
     io::{self, Write},
@@ -47,12 +47,12 @@ pub async fn run(
     };
 
     let colors = Vec::<Color>::deserialize(&mut socket).await?;
-    let mut app = Client::new_with_buffer(username.to_string(), initial_text, colors, Some(socket));
+    let mut app = App::new_with_buffer(username.to_string(), initial_text, colors, Some(socket));
 
     execute!(out, EnterAlternateScreen, EnableBracketedPaste)?;
     enable_raw_mode().unwrap();
 
-    app.redraw(&mut out)?;
+    app.client.redraw(&mut out)?;
 
     let mut reader = EventStream::new();
     loop {
@@ -77,7 +77,7 @@ pub async fn run(
                 })
             },
             length = async {
-                if let Some(x) = &mut app.buffers[app.current_buffer].socket{
+                if let Some(x) = &mut app.client.buffers[app.client.current_buffer].socket{
                     let mut buf = [0];
                     x.reader.peek(&mut buf).await
                 } else {
@@ -86,11 +86,11 @@ pub async fn run(
                 }
             } => {
                 assert_eq!(length?, 1, "The server disconnected");
-                app.curr_mut().update().await ?;
+                app.client.curr_mut().update().await ?;
                 Ok::<bool, io::Error>(true)
             },
             _ = async {
-                if let Some(timer) = app.modeinfo.timer.as_ref() {
+                if let Some(timer) = app.client.modeinfo.timer.as_ref() {
                     time::sleep_until(timer.deadline()).await;
                 } else {
                     future::pending::<()>().await;
@@ -104,8 +104,8 @@ pub async fn run(
             }
         }? {
             let size = terminal::size()?;
-            app.curr_mut().recalculate_cursor((size.0, size.1 - 1))?;
-            app.redraw(&mut out)?;
+            app.client.curr_mut().recalculate_cursor((size.0, size.1 - 1))?;
+            app.client.redraw(&mut out)?;
             out.flush()?;
         }
     }
